pyproject.toml
[project]
name = "webcourseslms"
version = "0.1.0"
description = ""
authors = [
    {name = "'alisheykhov99@gmail.com'"}
]
license = {text = "'MIT'"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
